Some Updates (Read them)
We've modified alot of things to perfect regeneration and weight now
affects running. There has been done some cleaning aswell. Enjoy.

- Ridiculous

+++ src/com/citelic/game/entity/npc/impl/familiar/Bullant.java
@@ -53,7 +53,7 @@ public class Bullant extends Familiar {
 			return false;
 		}
 		int agilityLevel = getOwner().getSkills().getLevel(Skills.AGILITY);
-		int runEnergy = player.getRunEnergy() + (Math.round(agilityLevel / 2));
+		double runEnergy = player.getRunEnergy() + (Math.round(agilityLevel / 2));
 		player.setNextGraphics(new Graphics(1300));
 		player.setNextAnimation(new Animation(7660));
 		player.setRunEnergy(runEnergy > 100 ? 100 : runEnergy);
